conditionally render tab component
