+++ react-app/src/components/UserPage/ProfileNav/AboutTab.js
@@ -1,0 +1,7 @@
+function AboutTab(){
+    return (
+        <h1>About Tab</h1>
+    )
+}
+
+export default AboutTab
