+++ react-app/src/components/UserPage/ProfileNav/AlbumsTab.js
@@ -1,0 +1,7 @@
+function AlbumsTab(){
+    return (
+        <h1>Albums Tab</h1>
+    )
+}
+
+export default AlbumsTab
