+++ react-app/src/components/UserPage/ProfileNav/PhotosTab.js
@@ -1,0 +1,7 @@
+function PhotosTab(){
+    return (
+        <h1>Photos Tab</h1>
+    )
+}
+
+export default PhotosTab
