+++ react-app/src/components/UserPage/ProfileNav/index.js
@@ -1,0 +1,29 @@
+import { useState } from "react";
+import AboutTab from "./AboutTab";
+import PhotosTab from "./PhotosTab";
+import AlbumsTab from "./AlbumsTab";
+
+function ProfileNav(){
+    const [selected, setSelected] = useState('')
+    return (
+        <>
+            <div className="profile-nav-bar">
+                <div onClick={() => setSelected('about')} className={selected === "about" ? "active-tab" : ""}>About</div>
+                <div onClick={() => setSelected('photos')} className={selected === "photos" ? "active-tab" : ""}>Photos</div>
+                <div onClick={() => setSelected('albums')} className={selected === "albums" ? "active-tab" : ""}>Albums</div>
+            </div>
+            {(selected === "about" &&
+                <AboutTab />
+            )}
+            {(selected === "photos" &&
+                <PhotosTab />
+            )}
+            {(selected === "albums" &&
+                <AlbumsTab />
+            )}
+
+        </>
+    )
+}
+
+export default ProfileNav
